refactor: Create index for ConnectionRequest folder
